Rename Types to Config, Move common out of config
Fix cabal file

Fix imports
