+++ keter.cabal
@@ -86,11 +86,11 @@ Library
     cpp-options:       -DSYSTEM_FILEPATH
 
   Exposed-Modules:     Keter.Plugin.Postgres
-                       Keter.Types
-                       Keter.Types.V04
-                       Keter.Types.V10
-                       Keter.Types.Common
-                       Keter.Types.Middleware
+                       Keter.Config
+                       Keter.Config.V04
+                       Keter.Config.V10
+                       Keter.Common
+                       Keter.Config.Middleware
                        Keter.App
                        Keter.AppManager
                        Keter.LabelMap
