+++ src/Keter/App.hs
@@ -15,13 +15,12 @@ module Keter.App
     , showApp
     ) where
 
-import Keter.Types.Common
+import Keter.Common
 import           Data.Set                   (Set)
 import           Data.Text                  (Text)
 import           System.FilePath            (FilePath)
 import           Data.Map                   (Map)
-import           Control.Exception          (SomeException)
-import           Keter.Rewrite (ReverseProxyConfig (..), RewriteRule (..))
+import           Keter.Rewrite (ReverseProxyConfig (..))
 import           Keter.TempTarball
 import           Control.Applicative       ((<$>), (<*>))
 import           Control.Arrow             ((***))
@@ -53,7 +52,7 @@ import           System.Directory          (canonicalizePath, doesFileExist,
                                             removeDirectoryRecursive)
 import           Keter.HostManager         hiding (start)
 import           Keter.PortPool            (PortPool, getPort, releasePort)
-import           Keter.Types
+import           Keter.Config
 import           Network.Socket
 import           Prelude                   hiding (FilePath)
 import           System.Environment        (getEnvironment)
@@ -62,7 +61,6 @@ import           System.Posix.Files        (fileAccess)
 import           System.Posix.Types        (EpochTime, GroupID, UserID)
 import           System.Timeout            (timeout)
 import qualified Network.TLS as TLS
-import qualified Data.Text.Encoding as Text
 
 data App = App
     { appModTime        :: !(TVar (Maybe EpochTime))
